HV-1667 Introduce a simplified ConstraintValidatorManager for predefined scope
