+++ engine/src/main/java/org/hibernate/validator/internal/engine/constraintvalidation/ConstraintValidatorManagerImpl.java
@@ -6,28 +6,20 @@
  */
 package org.hibernate.validator.internal.engine.constraintvalidation;
 
-import static org.hibernate.validator.internal.util.CollectionHelper.newArrayList;
-
 import java.lang.annotation.Annotation;
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 
-import javax.validation.ConstraintDeclarationException;
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorFactory;
-import javax.validation.metadata.ConstraintDescriptor;
 
-import org.hibernate.validator.constraintvalidation.HibernateConstraintValidator;
 import org.hibernate.validator.constraintvalidation.HibernateConstraintValidatorInitializationContext;
 import org.hibernate.validator.internal.metadata.descriptor.ConstraintDescriptorImpl;
 import org.hibernate.validator.internal.util.Contracts;
-import org.hibernate.validator.internal.util.TypeHelper;
 import org.hibernate.validator.internal.util.annotation.ConstraintAnnotationDescriptor;
 import org.hibernate.validator.internal.util.logging.Log;
 import org.hibernate.validator.internal.util.logging.LoggerFactory;
@@ -37,26 +29,9 @@ import org.hibernate.validator.internal.util.logging.LoggerFactory;
  *
  * @author Hardy Ferentschik
  */
-public class ConstraintValidatorManagerImpl implements ConstraintValidatorManager {
+public class ConstraintValidatorManagerImpl extends AbstractConstraintValidatorManagerImpl {
 
 	private static final Log LOG = LoggerFactory.make( MethodHandles.lookup() );
-
-	/**
-	 * The explicit or implicit default constraint validator factory. We always cache {@code ConstraintValidator}
-	 * instances if they are created via the default instance and with the default initialization context. Constraint
-	 * validator instances created via other factory instances (specified eg via {@code ValidatorFactory#usingContext()}
-	 * or initialization context are only cached for the most recently used factory and context.
-	 */
-	private final ConstraintValidatorFactory defaultConstraintValidatorFactory;
-
-	/**
-	 * The explicit or implicit default constraint validator initialization context. We always cache
-	 * {@code ConstraintValidator} instances if they are created via the default instance and with the default context.
-	 * Constraint validator instances created via other factory instances (specified eg via
-	 * {@code ValidatorFactory#usingContext()} or initialization context are only cached for the most recently used
-	 * factory and context.
-	 */
-	private final  HibernateConstraintValidatorInitializationContext defaultConstraintValidatorInitializationContext;
 
 	/**
 	 * The most recently used non default constraint validator factory.
@@ -88,8 +63,7 @@ public class ConstraintValidatorManagerImpl implements ConstraintValidatorManage
 	 */
 	public ConstraintValidatorManagerImpl(ConstraintValidatorFactory defaultConstraintValidatorFactory,
 			HibernateConstraintValidatorInitializationContext defaultConstraintValidatorInitializationContext) {
-		this.defaultConstraintValidatorFactory = defaultConstraintValidatorFactory;
-		this.defaultConstraintValidatorInitializationContext = defaultConstraintValidatorInitializationContext;
+		super( defaultConstraintValidatorFactory, defaultConstraintValidatorInitializationContext );
 		this.constraintValidatorCache = new ConcurrentHashMap<>();
 	}
 
@@ -133,9 +107,9 @@ public class ConstraintValidatorManagerImpl implements ConstraintValidatorManage
 	private <A extends Annotation> ConstraintValidator<A, ?> cacheValidator(CacheKey key,
 			ConstraintValidator<A, ?> constraintValidator) {
 		// we only cache constraint validator instances for the default and most recently used factory
-		if ( ( key.getConstraintValidatorFactory() != defaultConstraintValidatorFactory
+		if ( ( key.getConstraintValidatorFactory() != getDefaultConstraintValidatorFactory()
 				&& key.getConstraintValidatorFactory() != mostRecentlyUsedNonDefaultConstraintValidatorFactory ) ||
-				( key.getConstraintValidatorInitializationContext() != defaultConstraintValidatorInitializationContext
+				( key.getConstraintValidatorInitializationContext() != getDefaultConstraintValidatorInitializationContext()
 						&& key.getConstraintValidatorInitializationContext() != mostRecentlyUsedNonDefaultConstraintValidatorInitializationContext ) ) {
 
 			synchronized ( mostRecentlyUsedNonDefaultConstraintValidatorFactoryAndInitializationContextMutex ) {
@@ -196,102 +170,8 @@ public class ConstraintValidatorManagerImpl implements ConstraintValidatorManage
 		constraintValidatorCache.clear();
 	}
 
-	@Override
-	public ConstraintValidatorFactory getDefaultConstraintValidatorFactory() {
-		return defaultConstraintValidatorFactory;
-	}
-
-	@Override
-	public HibernateConstraintValidatorInitializationContext getDefaultConstraintValidatorInitializationContext() {
-		return defaultConstraintValidatorInitializationContext;
-	}
-
 	public int numberOfCachedConstraintValidatorInstances() {
 		return constraintValidatorCache.size();
-	}
-
-	/**
-	 * Runs the validator resolution algorithm.
-	 *
-	 * @param validatedValueType The type of the value to be validated (the type of the member/class the constraint was placed on).
-	 *
-	 * @return The class of a matching validator.
-	 */
-	private <A extends Annotation> ConstraintValidatorDescriptor<A> findMatchingValidatorDescriptor(ConstraintDescriptorImpl<A> descriptor, Type validatedValueType) {
-		Map<Type, ConstraintValidatorDescriptor<A>> availableValidatorDescriptors = TypeHelper.getValidatorTypes(
-				descriptor.getAnnotationType(),
-				descriptor.getMatchingConstraintValidatorDescriptors()
-		);
-
-		List<Type> discoveredSuitableTypes = findSuitableValidatorTypes( validatedValueType, availableValidatorDescriptors.keySet() );
-		resolveAssignableTypes( discoveredSuitableTypes );
-
-		if ( discoveredSuitableTypes.size() == 0 ) {
-			return null;
-		}
-
-		if ( discoveredSuitableTypes.size() > 1 ) {
-			throw LOG.getMoreThanOneValidatorFoundForTypeException( validatedValueType, discoveredSuitableTypes );
-		}
-
-		Type suitableType = discoveredSuitableTypes.get( 0 );
-		return availableValidatorDescriptors.get( suitableType );
-	}
-
-	private <A extends Annotation> List<Type> findSuitableValidatorTypes(Type type, Iterable<Type> availableValidatorTypes) {
-		List<Type> determinedSuitableTypes = newArrayList();
-		for ( Type validatorType : availableValidatorTypes ) {
-			if ( TypeHelper.isAssignable( validatorType, type )
-					&& !determinedSuitableTypes.contains( validatorType ) ) {
-				determinedSuitableTypes.add( validatorType );
-			}
-		}
-		return determinedSuitableTypes;
-	}
-
-	private <A extends Annotation> void initializeValidator(
-			ConstraintDescriptor<A> descriptor,
-			ConstraintValidator<A, ?> constraintValidator,
-			HibernateConstraintValidatorInitializationContext initializationContext) {
-		try {
-			if ( constraintValidator instanceof HibernateConstraintValidator ) {
-				( (HibernateConstraintValidator<A, ?>) constraintValidator ).initialize( descriptor, initializationContext );
-			}
-			constraintValidator.initialize( descriptor.getAnnotation() );
-		}
-		catch (RuntimeException e) {
-			if ( e instanceof ConstraintDeclarationException ) {
-				throw e;
-			}
-			throw LOG.getUnableToInitializeConstraintValidatorException( constraintValidator.getClass(), e );
-		}
-	}
-
-	/**
-	 * Tries to reduce all assignable classes down to a single class.
-	 *
-	 * @param assignableTypes The set of all classes which are assignable to the class of the value to be validated and
-	 * which are handled by at least one of the validators for the specified constraint.
-	 */
-	private void resolveAssignableTypes(List<Type> assignableTypes) {
-		if ( assignableTypes.size() == 0 || assignableTypes.size() == 1 ) {
-			return;
-		}
-
-		List<Type> typesToRemove = new ArrayList<>();
-		do {
-			typesToRemove.clear();
-			Type type = assignableTypes.get( 0 );
-			for ( int i = 1; i < assignableTypes.size(); i++ ) {
-				if ( TypeHelper.isAssignable( type, assignableTypes.get( i ) ) ) {
-					typesToRemove.add( type );
-				}
-				else if ( TypeHelper.isAssignable( assignableTypes.get( i ), type ) ) {
-					typesToRemove.add( assignableTypes.get( i ) );
-				}
-			}
-			assignableTypes.removeAll( typesToRemove );
-		} while ( typesToRemove.size() > 0 );
 	}
 
 	private static final class CacheKey {
